+++ flake.nix
@@ -57,6 +57,51 @@
               )
             '';
           };
+          # quarto: --syntax-highlighting
+          pandoc = pkgs.stdenv.mkDerivation {
+            pname = "p-pandoc";
+            version = "3.9.0.2";
+            src = pkgs.fetchurl {
+              url = "https://github.com/jgm/pandoc/releases/download/3.9.0.2/pandoc-3.9.0.2-linux-amd64.tar.gz";
+              sha256 = "sha256-ppq/q6vailaWmiVLCflVOnvond7ADU4P6f1YXXGmdQg=";
+            };
+
+            dontConfigure = true;
+            dontBuild = true;
+
+            installPhase = ''
+              runHook preInstall
+              install -Dm755 -t $out/bin bin/pandoc bin/pandoc-lua bin/pandoc-server
+              install -Dm644 -t $out/share/man/man1 share/man/man1/*.1.gz
+              runHook postInstall
+            '';
+
+            meta = {
+              homepage = "https://pandoc.org";
+              description = "Universal markup converter (upstream prebuilt static binary)";
+              mainProgram = "pandoc";
+              license = pkgs.lib.licenses.gpl2Plus;
+              platforms = [ "x86_64-linux" ];
+            };
+          };
+          pythonPackages =  ps: with ps; [
+                  jupyter
+                  ipykernel
+                  nbclient
+                  nbformat
+                  pandas
+                  tabulate
+                  polars
+                  altair
+                  vl-convert-python
+                ];
+          python = pkgs.python3.withPackages pythonPackages;
+          quarto = (
+            pkgs.quarto.override {
+              inherit pandoc;
+              extraPythonPackages = pythonPackages;
+            }
+          );
         in
         with pkgs;
         {
@@ -85,7 +130,21 @@
               }
             ];
             language.rust.enableDefaultToolchain = false;
-            packages = [ toolchain ff ];
+            packages = [
+              toolchain
+              gnumake
+              ff
+
+              gdb
+              # benchmarks
+              heaptrack
+              perf
+              wrk
+              wrk2
+              xan
+              quarto
+              python
+            ];
           };
         };
     };